Add new applications
The game server added more apps, so I added them to the list. I also added a thing to automatically update the list from the game server if possible.

Since the endpoint 401s if you aren't in the server, a user must be in the game server to update the list, so I added a backup list. But if that list gets out of date, there could be problems, so I added a toast to notify any interested maintainers that the backup list needs to be updated.